Emails: Move IntervalLength type and function to new folder
