+++ client/my-sites/email/email-providers-comparison/billing-interval-toggle/index.tsx
@@ -1,6 +1,6 @@
 import { useTranslate } from 'i18n-calypso';
 import SegmentedControl from 'calypso/components/segmented-control';
-import { IntervalLength } from 'calypso/my-sites/email/email-providers-stacked-comparison/provider-cards/utils';
+import { IntervalLength } from 'calypso/my-sites/email/email-providers-comparison/interval-length';
 import type { ReactElement } from 'react';
 
 import './style.scss';
