+++ client/my-sites/email/email-providers-stacked-comparison/provider-cards/utils.tsx
@@ -7,23 +7,6 @@ import type {
 	MinimalRequestCartProduct,
 } from '@automattic/shopping-cart';
 
-export enum IntervalLength {
-	ANNUALLY = 'annually',
-	MONTHLY = 'monthly',
-}
-
-export const castIntervalLength = ( stringValue: string ): IntervalLength | undefined => {
-	const isStringValueInEnum = Object.values( IntervalLength ).some(
-		( enumValue: string ) => enumValue === stringValue
-	);
-
-	if ( isStringValueInEnum ) {
-		return stringValue as IntervalLength;
-	}
-
-	return undefined;
-};
-
 export const addToCartAndCheckout = (
 	shoppingCartManager: ShoppingCartManagerActions,
 	cartItem: RequestCartProduct | MinimalRequestCartProduct,
